Bookmark triangles are drawn at zero position after any of them is removed. This happens due to failing conversion from QStrings containing non-integers to int.

+++ src/toolbar/MainToolbar.cpp
@@ -571,7 +571,7 @@ MainToolbar::updateBookmarks( const QString *BookmarkName )
             {
                 if ( url->command() == "play" && url->args().keys().contains( "pos" ) )
                 {
-                    int pos = url->args().value( "pos" ).toInt() * 1000;
+                    int pos = url->args().value( "pos" ).toDouble() * 1000;
                     debug() << "showing timecode: " << url->name() << " at " << pos ;
                     m_slider->drawTriangle( url->name(), pos, ( BookmarkName && BookmarkName == url->name() ) );
                 }
